chore(deps): bump github.com/anchore/go-collections and rollback getProviders
