+++ syft/get_source_config.go
@@ -2,7 +2,9 @@ package syft
 
 import (
 	"crypto"
+	"fmt"
 
+	"github.com/anchore/go-collections"
 	"github.com/anchore/stereoscope/pkg/image"
 	"github.com/anchore/syft/syft/source"
 	"github.com/anchore/syft/syft/source/sourceproviders"
@@ -57,6 +59,29 @@ func (c *GetSourceConfig) WithSources(sources ...string) *GetSourceConfig {
 func (c *GetSourceConfig) WithDefaultImagePullSource(defaultImagePullSource string) *GetSourceConfig {
 	c.DefaultImagePullSource = defaultImagePullSource
 	return c
+}
+
+func (c *GetSourceConfig) getProviders(userInput string) ([]source.Provider, error) {
+	providers := collections.TaggedValueSet[source.Provider]{}.Join(sourceproviders.All(userInput, c.SourceProviderConfig)...)
+
+	// if the "default image pull source" is set, we move this as the first pull source
+	if c.DefaultImagePullSource != "" {
+		base := providers.Remove(sourceproviders.PullTag)
+		pull := providers.Select(sourceproviders.PullTag)
+		def := pull.Select(c.DefaultImagePullSource)
+		if len(def) == 0 {
+			return nil, fmt.Errorf("invalid DefaultImagePullSource: %s; available values are: %v", c.DefaultImagePullSource, pull.Tags())
+		}
+		providers = base.Join(def...).Join(pull...)
+	}
+
+	// narrow the sources to those explicitly requested generally by a user
+	if len(c.Sources) > 0 {
+		// select the explicitly provided sources, in order
+		providers = providers.Select(c.Sources...)
+	}
+
+	return providers.Values(), nil
 }
 
 func DefaultGetSourceConfig() *GetSourceConfig {
