+++ syft/get_source_test.go
@@ -3,7 +3,6 @@ package syft
 import (
 	"testing"
 
-	"github.com/anchore/stereoscope"
 	"github.com/stretchr/testify/require"
 
 	"github.com/anchore/stereoscope/pkg/image"
@@ -19,35 +18,6 @@ type mockSource struct {
 
 func (s mockSource) Describe() source.Description {
 	return s.desc
-}
-
-func TestGetProviders_DefaultImagePullSource(t *testing.T) {
-	userInput := ""
-	cfg := &GetSourceConfig{DefaultImagePullSource: stereoscope.RegistryTag}
-	allSourceProviders := sourceproviders.All(userInput, cfg.SourceProviderConfig)
-
-	providers, err := getProviders(userInput, cfg)
-	if err != nil {
-		t.Errorf("Expected no error for DefaultImagePullSource parameter, got: %v", err)
-	}
-
-	if len(providers) != len(allSourceProviders) {
-		t.Errorf("Expected %d providers, got %d", len(allSourceProviders), len(providers))
-	}
-}
-
-func TestGetProviders_Sources(t *testing.T) {
-	userInput := ""
-	cfg := &GetSourceConfig{Sources: []string{stereoscope.RegistryTag}}
-
-	providers, err := getProviders(userInput, cfg)
-	if err != nil {
-		t.Errorf("Expected no error for Sources parameter, got: %v", err)
-	}
-
-	if len(providers) != 1 {
-		t.Errorf("Expected 1 providers, got %d", len(providers))
-	}
 }
 
 func TestValidateSourcePlatform_NilSource(t *testing.T) {
